Improve availability handling
Update handling of the available field to correspond to API changes (it now reflects the total available to the user, ignoring anything they might have already ordered).

Change the availability indication so we don't show the quantity available, but if they try to order more than are available, show an error.