Change google auth/config mechanism
For some reason the previous method of JWT authentication isn't working with newly-created service accounts. I can't really find any documentation for it, so maybe it's deprecated or something. The new method requires a path to a file instead of in-memory config, which is annoying, but doable with a little fiddling.