Improve order page UI
Change the order page UI to hopefully be more usable. Make the images bigger and make the controls an overlay on top of them to look less 1995. Make the primary way of ordering a select control that saves the order when it's changed (so the user doesn't have to click a submit button that everyone always forgets). Then make a secondary way of ordering a quantity bigger than the options in the select allow that is a number input with a save button, but that also auto-saves as soon as it's blurred.

This definitely looks nicer, and hopefully is easier for users to operate!